+++ packages/component-library/src/components.d.ts
@@ -384,6 +384,10 @@ export namespace Components {
     }
     interface PlmgTextInput {
         /**
+          * Define disabled  Allowed value: boolean  Disables text input
+         */
+        "disabled": boolean;
+        /**
           * Define error message  Allowed value: any string  Sets error style and error message
          */
         "errorMessage": string;
@@ -391,6 +395,10 @@ export namespace Components {
           * Define a label name for the input field.  Allowed values: - Any string  A unique label name for each element in a form is required for accessibility
          */
         "label": string;
+        /**
+          * Define readonly  Allowed value: boolean  Makes text input read only
+         */
+        "readOnly": boolean;
         /**
           * Define if an input is required.  Allowed values: - true - false  Default: false
          */
@@ -407,6 +415,10 @@ export namespace Components {
           * Define tip  Allowed value: any string  Displays a tip message
          */
         "tip": string;
+        /**
+          * Control the text input's value  Allowed values: - Any string  Sets the value of the text input
+         */
+        "value": string;
     }
     interface PlmgTooltip {
         /**
@@ -994,6 +1006,10 @@ declare namespace LocalJSX {
     }
     interface PlmgTextInput {
         /**
+          * Define disabled  Allowed value: boolean  Disables text input
+         */
+        "disabled"?: boolean;
+        /**
           * Define error message  Allowed value: any string  Sets error style and error message
          */
         "errorMessage"?: string;
@@ -1005,6 +1021,10 @@ declare namespace LocalJSX {
           * Event emitted when value changed
          */
         "onValueUpdated"?: (event: PlmgTextInputCustomEvent<any>) => void;
+        /**
+          * Define readonly  Allowed value: boolean  Makes text input read only
+         */
+        "readOnly"?: boolean;
         /**
           * Define if an input is required.  Allowed values: - true - false  Default: false
          */
@@ -1021,6 +1041,10 @@ declare namespace LocalJSX {
           * Define tip  Allowed value: any string  Displays a tip message
          */
         "tip"?: string;
+        /**
+          * Control the text input's value  Allowed values: - Any string  Sets the value of the text input
+         */
+        "value"?: string;
     }
     interface PlmgTooltip {
         /**