+++ packages/component-library/src/components/plmg-text-input/plmg-text-input.stories.js
@@ -6,6 +6,11 @@ export default {
   parameters: {},
   decorators: [],
   argTypes: {
+    disabled: {
+      control: {
+        type: 'boolean',
+      },
+    },
     ['error-message']: {
       control: { type: 'text' },
     },
@@ -14,6 +19,11 @@ export default {
     },
     ['show-label']: {
       control: { type: 'boolean' },
+    },
+    ['read-only']: {
+      control: {
+        type: 'boolean',
+      },
     },
     required: {
       control: { type: 'boolean' },
@@ -25,16 +35,22 @@ export default {
     tip: {
       control: { type: 'text' },
     },
+    value: {
+      control: { type: 'text' },
+    },
   },
 };
 
 const PROPS = [
+  'disabled',
   'error-message',
   'show-label',
   'label',
+  'read-only',
   'required',
   'size',
   'tip',
+  'value',
 ];
 
 const EVENTS = [];
@@ -54,6 +70,7 @@ Primary.args = {
   ['error-message']: '',
   ['show-label']: false,
   required: false,
+  value: '',
 };
 
 export const AllSizes = (args) => {
@@ -73,6 +90,21 @@ export const AllSizes = (args) => {
   return el;
 };
 AllSizes.storyName = 'All Sizes';
+
+export const disabled = Template.bind({});
+disabled.storyName = 'Disabled';
+disabled.args = {
+  label: 'Disabled',
+  disabled: true,
+};
+
+export const readOnly = Template.bind({});
+readOnly.storyName = 'Read Only';
+readOnly.args = {
+  label: 'Read Only',
+  ['read-only']: true,
+  value: 'Cannot be edited, but can be copied',
+};
 
 export const showLabel = Template.bind({});
 showLabel.storyName = 'Label';